# Attaches row to DOM after all columns have been added