--- a/grunt/config/jshint.js
+++ b/grunt/config/jshint.js
@@ -1,7 +1,8 @@
 module.exports = {
 	options: {
-		jshintrc: '.jshintrc' // use project defined jshint settings which can be shared with IDEs etc
+		jshintrc: '.jshintrc', // use project defined jshint settings which can be shared with IDEs etc
+		ignores: ['test/data/*.js']
 	},
 	sourceAndDist: ['Gruntfile.js', 'js/*.js', 'dist/fuelux.js'],
-	tests: ['test/**/*.js', '!test/commonjs-bundle.js']
+	tests: ['test/**/*.js', '!test/commonjs-bundle.js'],
 };