--- a/package.json
+++ b/package.json
@@ -19,7 +19,7 @@
     "eslint": "^3.3.1",
     "eslint-config-airbnb-es5": "^1.0.9",
     "eslint-plugin-requirejs": "^2.0.1",
-    "grunt": "0.4.5",
+    "grunt": "^0.4.5",
     "grunt-banner": "0.6.0",
     "grunt-blanket-qunit": "0.2.0",
     "grunt-browserify": "5.0.0",