@interactivellama 2016-09-12
I can confirm about a half second improvement in FireFox.