@interactivellama 2016-09-12
Looks good to me. There are so many variations for Repeater, it's difficult to review all of them, but I've read every line, confirmed that this is faster and was unable to find any bugs in the repeater examples I have.